+++ packages/cms/migrations/20250429113608_add_topic_related_twreporter_articles/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE `Topic` ADD COLUMN `relatedTwreporterArticles` JSON NULL;