+++ packages/frontend/src/fetchers/party.ts
@@ -1,6 +1,8 @@
+'use client'
 import useSWR from 'swr'
 
 export type partyData = {
+  id: number
   slug: string
   name: string
   imageLink?: string